Pass the payment id along during the edd_commission_info filter
I have a custom EDD payment meta field that records the gateway fee for a transaction. I want to take this fee into consideration when making payouts to my users, so that I'm not paying out portions of payments that I never actually received.

This filter allows for me to look up a payment's metadata, then use it however I desire in order to make modifications to the commission info.