Improper Download Price Option name #96
This fixes a bug noted #96 where the creation of recorded commissions when a variable price is added to the cart first, and then a single price, the variable Price ID name was added to the single price post meta.

It prevents this recording from happening here on out, and fixes old data by making sure the commission download has variable prices before showing the price opiton name in the list table.

Also prevents commissions for deleted users from throwing a ntoice on the list table.