Add Commission Shipping Fee controls per product
This adds a "Shipping" control to the "Commissions" metabox which allows the site owner to set how the shipping fee shoud be split based on the product. This will allow a certian product to have the shipping fee paid to a specific user (the person shipping the item), give the shipping fee to the site, or to split the shipping fee among all getting comissions based on their commission amount.