Move all Shipping calculation functions
In order to keep the code organized, this makes it possible to have all shipping-related fee calculations for comissions be in their own file. It adds a few extra args to the eddc_calc_commission_amount function including recipient counter (so you know whether it is the first recipient or not), and the cart_item object. This commit also changes the variable $download to the more accurately-names “cart_item” as it is a cart item we are looping through, not a typical “download”.